+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+- Adds [`Range.prototype.isUnbounded`][].
+
+[`Range.prototype.isUnbounded`]: https://github.com/moll/js-strange/blob/master/doc/API.md#Range.prototype.isUnbounded
+
 ## 1.1.0 (Jun 29, 2015)
 - Adds `Range.prototype.isEmpty`.
 - Adds `Range.prototype.intersects`.